% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARtest.R
\name{DItest}
\alias{DItest}
\title{Perform the over(under)-Dispersion Index test for the INAR(1) (Weiss et al, 2019)}
\usage{
DItest(X)
}
\arguments{
\item{X}{vector, a series of counts.}
}
\value{
........
}
\description{
Perform the over(under)-Dispersion Index test for the INAR(1) (Weiss et al, 2019)
}
\details{
The function performs the zero inflation test for the INAR(1) model.
}
\examples{
# ....... examples .....
# DItest(X = rpois(100,2))

}
