% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katz.R
\name{dkatz}
\alias{dkatz}
\title{Katz probability mass function}
\usage{
dkatz(x, a, b, log = FALSE, tol = 1e-10)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{a}{Katz parameter, a >= 0.}

\item{b}{Katz parameter, b < 1.}

\item{log}{Logical; if TRUE, probabilities are returned on log scale.}

\item{tol}{Numerical tolerance.}
}
\description{
Katz probability mass function
}
\details{
The Katz family is defined by Sun and McCabe as:

$$
p_{j+1} / p_j = (a + b j) / (j + 1), j = 0, 1, 2, ...
$$

with a > 0 and b < 1.

Special cases:
\itemize{
\item b < 0: Binomial, with m = -a / b and p = b / (b - 1);
\item b = 0: Poisson, with lambda = a;
\item 0 < b < 1: Negative Binomial, with r = a / b and p = b.
}
}
