% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\docType{data}
\name{downloads}
\alias{downloads}
\title{Download counts}
\format{
An object of class \code{"tibble"};
}
\source{
Wiley repository, <www.wiley.com/go/weiss/discrete-valuedtimeseries>
}
\usage{
data(downloads)
}
\description{
This data contain a time series expressing the daily number of
downloads of a TEX editor for the period from June 2006 to
February 2007 (T = 267)
}
\examples{
data(downloads)
\donttest{
plot(downloads$X,type="b")}
}
\references{
Weiß, C. H. (2018). An introduction to discrete-valued time
series. John Wiley & Sons.

Weiß, C. H. (2008). Thinning operations for modeling time series
of counts—a survey. AStA Advances in Statistical Analysis, 92(3), 319-341.
}
\keyword{datasets}
