% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ecdfcpp}
\alias{ecdfcpp}
\title{Compute the empirical cumulative distribution function of a numeric vector}
\usage{
ecdfcpp(eval, x)
}
\arguments{
\item{eval}{NumericVector}

\item{x}{NumericVector}
}
\description{
Compute the empirical cumulative distribution function of a numeric vector
}
\details{
This is an internal function, it will be excluded in future versions.
}
