% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARfit.R
\name{est_mom}
\alias{est_mom}
\title{Estimation of the innovation process' parameters}
\usage{
est_mom(mX, varX, mINN, varINN, arrival)
}
\arguments{
\item{mX}{mean of the INAR process}

\item{mINN}{mean of the innovation process}

\item{arrival}{distribution of the innovation process}

\item{vX}{variance of the INAR process}

\item{vINN}{variance of the innovation process}
}
\value{
A list with parameter estimates
}
\description{
Estimation of the innovation process' parameters
}
\details{
Inner function that estimates the parameters related with the innovation process, given anINAR(p) model.
}
