% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINAR.R
\name{estimMYW}
\alias{estimMYW}
\title{Yule-Walker for MINAR(p) alphas parameter estimation}
\usage{
estimMYW(X, p, inn = "poi", ...)
}
\arguments{
\item{X, }{matrix of m observed series}

\item{p, }{number of lags}

\item{inn, }{distribution of the innovation process}

\item{..., }{additional parameters}
}
\description{
Experimental function
}
\examples{
# example

}
