% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARgen.R
\name{genINAR}
\alias{genINAR}
\title{Generate INAR(p) models with different innovations}
\usage{
genINAR(n, a, par = NULL, inn = "poi", custominn = NULL, burnout = 500)
}
\arguments{
\item{n}{integer, number of observations, length of the series.}

\item{a}{vector, thinning parameters. The lenght of this vector defines the number of lags \code{p} of the INAR(p) process.}

\item{par}{vector, parameters related with the model, see the details section.}

\item{inn}{character, innovation distribution. Default value is \code{"poi"}, alternative values are \code{"negbin"} for Negative Binomial, \code{"genpoi"} for Generalized Poisson and \code{"katz"} for the Katz family.}

\item{custominn}{list, a custom function to generate innovations. If provided, this function will be used instead of the built-in ones. The list should contain two elements: \code{dgp}, a function that generates the innovations and \code{par}, a list of parameters to be passed to the dgp function. The dgp function should take an integer n as its first argument and return a numeric vector of length n containing the generated innovations. The par list can contain any additional parameters required by the dgp function.}

\item{burnout}{integer, number of starting observations to discard. Set to 500 by default.}
}
\value{
A number.
}
\description{
Generate INAR(p) models with different innovations
}
\details{
The function generates \code{n} observations drawn from an INAR(p) model
\deqn{X_t = \alpha_1 {*} X_{t-1} + \ldots + \alpha_p {*} X_{t-p} + \varepsilon_t}
where \eqn{*} is the binomial thinning operator and \eqn{\alpha_p} are the thinning parameters \insertCite{Steutel1979,al1987first}{INAr}.
The values of the thinning parameters \eqn{\alpha_i} are the elements of the \code{a} vector, for \eqn{i=1, \ldots, p}, while in \code{par} vector are specified the parameters related with the innovation distribution \verb{\\varepsilon}. The available processes for the innovations are listed below.

\strong{Poisson}. Innovations are drawn from a Poisson distribution \eqn{Poi(\lambda)}
\deqn{P(\varepsilon_t = k)= \dfrac{\lambda^k e^{-\lambda}}{k!},}
for \eqn{k \in \mathbb{N}}. The positive real number \eqn{\lambda} is equal to the expected value and also to the variance. The parameter is then equal to:
\itemize{
\item \code{par} = c(\eqn{\lambda})
}

\strong{Negative Binomial}. Innovations are drawn from a Negative Binomial distribution \eqn{Poi(\lambda)}
\deqn{P(\varepsilon_t = k) = \dfrac{\Gamma(k+\gamma)}{\Gamma(\gamma)k!}p^k (1-p)^\gamma ,}
for \eqn{k \in \mathbb{N}}. The positive real number \eqn{\gamma} is the size and \eqn{p \in (0,1)} is the probability. The parameter is then equal to:
\itemize{
\item \code{par} = c(\eqn{\gamma,p})
}

\strong{Generalized Poisson}. Innovations are \eqn{\varepsilon_t \sim Poi(\lambda)}
\itemize{
\item \code{par} = c(\eqn{\lambda,\kappa}) ...
}

\strong{Katz}. Innovations are \eqn{\varepsilon_t \sim Poi(\lambda)}
\itemize{
\item \code{par} =  c(\eqn{\alpha,\beta}).. alpha and beta ...
}

\strong{Zero-Inflated Poisson} \insertCite{piancastelli2019inferential}{INAr}. Innovations are \eqn{\varepsilon_t \sim ZIP(\lambda,\rho)}
\deqn{P(\varepsilon_t = k) = \begin{cases}
\rho + (1-\rho) e^{-\lambda}, & \text{if } k = 0, \\
(1-\rho)\dfrac{e^{-\lambda} \lambda^k}{k!}, & \text{if } k > 0,
\end{cases}}
for \eqn{k \in \mathbb{N}}, where \eqn{\lambda \geq 0} is the Poisson parameter and \eqn{\rho \in [0,1]} is the zero-inflation mixing parameter.
\itemize{
\item \code{par} =  c(\eqn{\lambda,\rho}).
}

By the way, gatto gatto miao miao \code{\link[=genINAR]{genINAR()}}.
}
\examples{
# Generate 500 observations from an INAR(1) model with
# thinning parameter \eqn{\alpha}=0.5 and Poisson(2) innovations
set.seed(1234)
lam <- 2
genINAR(500, a = 0.5, par = lam, inn = "poi")

}
\references{
\insertAllCited{}
}
