% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINAR.R
\name{genMINAR}
\alias{genMINAR}
\title{Generate MINAR(p) models with different innovations}
\usage{
genMINAR(n, m, A_list, lambdas, inn = "poi", burnout = 500)
}
\arguments{
\item{n}{integer, number of observations, length of the series.}

\item{m}{integer, number of series, dimension of the process.}

\item{A_list}{vector, thinning parameters. The lenght of this vector defines the number of lags \code{p} of the INAR(p) process.}

\item{lambdas}{vector, parameters of the innovation distribution.}

\item{inn}{character, innovation distribution. Default value is \code{"poi"}, alternative values are \code{"negbin"} for Negative Binomial, \code{"genpoi"} for Generalized Poisson and \code{"katz"} for the Katz family.}

\item{burnout}{integer, number of starting observations to discard. Set to 500 by default.}
}
\description{
Experimental function
}
\examples{
# example
}
