% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARviz.R
\name{ggacf}
\alias{ggacf}
\title{ACF plot for INAR data}
\usage{
ggacf(x, lag.max = NULL, plot = TRUE, ...)
}
\arguments{
\item{x, }{observed series}

\item{lag.max, }{maximum lag to display}

\item{plot, }{if TRUE, returns a ggplot object; if FALSE, returns
the ACF object}

\item{..., }{additional options}
}
\value{
A ggplot object if \code{plot = TRUE}; otherwise an object of class \code{"acf"}.
}
\description{
Plot the autocorrelation function (ACF) for INAR data. This is a modified version of the \code{pacf} function from the package \code{stats}, adapted to handle INAR data and to return a ggplot object for visualization.
}
