% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HMC_BOOT_Cpp}
\alias{HMC_BOOT_Cpp}
\title{Bootstrap version of the Harris-McCabe score test.
!!!WARNING!!! Still under development, do not use! It will be replaced by INARtest() in future versions.}
\usage{
HMC_BOOT_Cpp(x, B)
}
\arguments{
\item{x}{NumericVector}

\item{B}{int}
}
\description{
Bootstrap version of the Harris-McCabe score test.
!!!WARNING!!! Still under development, do not use! It will be replaced by INARtest() in future versions.
}
\details{
This is an internal function, it will be excluded in future versions.
}
