% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HMC_Cpp}
\alias{HMC_Cpp}
\title{Harris-McCabe score statistic to test for dependence in an integer autoregressive process}
\usage{
HMC_Cpp(x)
}
\arguments{
\item{x}{NumericVector}
}
\description{
Harris-McCabe score statistic to test for dependence in an integer autoregressive process
}
\details{
This is an internal function, it will be excluded in future versions.
}
