% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARtest.R
\name{HMCtest}
\alias{HMCtest}
\title{Perform Harris-McCabe INAR(1) test.}
\usage{
HMCtest(X, B = 0, saveboot = FALSE)
}
\arguments{
\item{X}{vector, the observed series of counts.}

\item{B}{integer, the number of bootstrap samples to be generated.}

\item{saveboot}{logical, if TRUE the bootstrap replicates are saved in the output object (only if B > 0).
The default value is 0, corresponding to no bootstrap samples.}
}
\value{
A number.
}
\description{
Perform Harris-McCabe INAR(1) test.
}
\details{
The function performs the Harris-McCabe (HMC) tests for the INAR
model. It is possible to perform the exact test using the original data (B=0)
or using bootstrap samples (B > 0).
}
\examples{
# ....... examples .....
# HMCtest(X = rpois(100,2), B = 0)

}
