% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HMCtest_boot}
\alias{HMCtest_boot}
\title{Wrapper function for compution the Harris-McCabe bootstrap score test.}
\usage{
HMCtest_boot(X, B)
}
\arguments{
\item{B}{int}

\item{x}{NumericVector}
}
\description{
Wrapper function for compution the Harris-McCabe bootstrap score test.
}
\details{
This is an internal function, it will be excluded in future versions.
}
