% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARfit.R, R/genericfuns.R
\name{INAR}
\alias{INAR}
\alias{summary.INAR}
\alias{fitted.INAR}
\alias{predict.INAR}
\alias{plot.INAR}
\title{Fitting INAR(p) Models}
\usage{
INAR(X, p, inn = "poi", method = "CLS")

\method{summary}{INAR}(object, ...)

\method{fitted}{INAR}(object, ...)

\method{predict}{INAR}(
  object,
  n.ahead = 1,
  type = c("mean", "bootstrap"),
  B = 999,
  level = 0.95,
  seed = NULL,
  ...
)

\method{plot}{INAR}(x, ...)
}
\arguments{
\item{X}{data vector}

\item{p}{the order of the INAR(p) process}

\item{inn}{distribution of the innovation process, one of "poi" (Poisson), "negbin" (Negative Binomial), "genpoi" (Generalized Poisson), "katz" (Katz)}

\item{method}{estimation method, one of "YW" (Yule-Walker), "CLS" (Conditional Least Squares), "CML" (Conditional Maximum Likelihood), "SP" (Saddlepoint Approximation)}

\item{object, }{an \code{INAR} object}

\item{...}{further arguments.}

\item{n.ahead}{forecast horizon.}

\item{type}{either "mean" or "bootstrap".}

\item{B}{number of bootstrap trajectories.}

\item{level}{prediction interval level.}

\item{seed}{optional random seed.}

\item{x, }{an \code{INAR} object}

\item{object}{an object of class "INAR".}

\item{..., }{additional options}
}
\value{
The fitted model, an object of class \code{INAR}

an object of class "INARforecast".

the original object, invisibly.
}
\description{
summary method for class \code{INAR}.

summary method for class \code{INAR}.
}
\details{
This function estimates an INAR(p) model given the distribution of the innovation process.
}
