% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARfit.R
\name{INARfit}
\alias{INARfit}
\title{Estimation of the innovation process' parameters}
\usage{
INARfit(X, order, arrival = "poisson")
}
\arguments{
\item{X}{data vector}

\item{order}{p, the order of the INAR(p) process}

\item{arrival}{distribution of the innovation process}
}
\value{
The fitted model
}
\description{
Estimation of the innovation process' parameters
}
\details{
Inner function that estimates the parameters related with the innovation process, given anINAR(p) model.
}
