% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericfuns.R
\name{print.INARtest}
\alias{print.INARtest}
\title{Print the results of INAR tests.}
\usage{
\method{print}{INARtest}(x, digits = getOption("digits"), prefix = "\\t", ...)
}
\arguments{
\item{x, }{an \code{INARtest} object.}

\item{digits, }{number of digits.}

\item{prefix, }{prefix for the output.}

\item{..., }{additional options.}
}
\description{
This is a modified version of the print function for class \code{htest} from the package \code{stats}.
}
