% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARtest.R
\name{IPtest}
\alias{IPtest}
\title{Poissonity test for the presence of P-INAR(1) models based on the comparison of the  parametric and semiparametric bootstrap distributions of the SMC statistic.}
\usage{
IPtest(X, B1 = 399, B2 = 399, inn = "poi", saveboot = FALSE)
}
\arguments{
\item{X}{vector, a series.}

\item{B1}{integer, the number of bootstrap samples to be generated in the first step of the test.}

\item{B2}{integer, the number of bootstrap samples to be generated in the second}

\item{inn}{character, the innovation distribution to be used in the test. The options are "poi", "negbin" and "genpoi".}

\item{saveboot}{logical, if TRUE the bootstrap replicates are saved in the output object (only if B1 > 0 and B2 > 0).}
}
\value{
........
}
\description{
Poissonity test for the presence of P-INAR(1) models based on the comparison of the  parametric and semiparametric bootstrap distributions of the SMC statistic.
}
\details{
This function performs the Sun-McCabe (SMC) test for the presence of P-INAR(1) models.
The test is based on a double bootstrap procedure, where:
\itemize{
\item in the first step B1 bootstrap samples are generated under the null hypothesis of a P-INAR(1) model,
\item in the second step B2 bootstrap samples are generated under the alternative hypothesis of a general INAR(1) model.
The test statistic is computed performing KS tests between the distribution of the test statistic under the null and the alternative hypothesis, obtained from the two bootstrap steps.
}
}
\examples{
# ....... examples .....
# IPtest(X = rpois(100,2), B1 = 399, B2 = 99, inn = "poi")

}
