% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MINARp_gen_cpp}
\alias{MINARp_gen_cpp}
\title{Generate a MINAR(p) series}
\usage{
MINARp_gen_cpp(resid, A)
}
\arguments{
\item{resid}{Numeric matrix of size n x m, where n is the number of time points and m is the number of series.}

\item{A}{Numeric matrix of size m x m*p, where m is the number of series. Each column of A contains the autoregressive parameters for the corresponding lag.}
}
\description{
Experimental function
}
\details{
This is an internal function.
}
