% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katz.R
\name{pkatz}
\alias{pkatz}
\title{Katz cumulative distribution function}
\usage{
pkatz(q, a, b, lower.tail = TRUE, log.p = FALSE, tol = 1e-10)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{a}{Katz parameter.}

\item{b}{Katz parameter.}

\item{lower.tail}{Logical; if TRUE, returns P(X <= q).}

\item{log.p}{Logical; if TRUE, probabilities are returned on log scale.}

\item{tol}{Numerical tolerance.}
}
\description{
Katz cumulative distribution function
}
