% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericfuns.R
\name{plot.INARtest}
\alias{plot.INARtest}
\title{Plot bootstrap distribution of an INAR test}
\usage{
\method{plot}{INARtest}(x, which = "density", ...)
}
\arguments{
\item{x}{an object of class \code{INARtest}.}

\item{which}{a character string specifying the type of plot to produce. Options are "density" for a density plot or "ecdf" for an empirical cumulative distribution function (ECDF) plot.}

\item{...}{additional arguments.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot bootstrap distribution of an INAR test
}
