% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katz.R
\name{qkatz}
\alias{qkatz}
\title{Katz quantile function}
\usage{
qkatz(p, a, b, lower.tail = TRUE, log.p = FALSE, tol = 1e-10)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{a}{Katz parameter.}

\item{b}{Katz parameter.}

\item{lower.tail}{Logical; if TRUE, probabilities are P(X <= x).}

\item{log.p}{Logical; if TRUE, probabilities are supplied on log scale.}

\item{tol}{Numerical tolerance.}
}
\description{
Katz quantile function
}
