% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RHO_BOOT_Cpp}
\alias{RHO_BOOT_Cpp}
\title{Autocorrelation bootstrap test.}
\usage{
RHO_BOOT_Cpp(x, B)
}
\arguments{
\item{x}{NumericVector}

\item{B}{int}
}
\description{
Autocorrelation bootstrap test.
}
\details{
This is an internal function, it will be excluded in future versions.
}
