% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RHO_Cpp}
\alias{RHO_Cpp}
\title{Autocorrelation rho test statistic to test for unit root in an integer autoregressive process}
\usage{
RHO_Cpp(x)
}
\arguments{
\item{x}{NumericVector}
}
\description{
Autocorrelation rho test statistic to test for unit root in an integer autoregressive process
}
\details{
This is an internal function, it will be excluded in future versions.
}
