% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katz.R
\name{rkatz}
\alias{rkatz}
\title{Random generation from the Katz distribution}
\usage{
rkatz(n, a, b, tol = 1e-10)
}
\arguments{
\item{n}{Number of observations.}

\item{a}{Katz parameter.}

\item{b}{Katz parameter.}

\item{tol}{Numerical tolerance.}
}
\description{
Random generation from the Katz distribution
}
