% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SMC_pitBOOT_Cpp}
\alias{SMC_pitBOOT_Cpp}
\title{PIT bootstrap version of the Sun-McCabe score test.}
\usage{
SMC_pitBOOT_Cpp(x, B, method)
}
\arguments{
\item{x}{NumericVector}

\item{B}{int}

\item{method}{unsigned int}
}
\description{
PIT bootstrap version of the Sun-McCabe score test.
}
\details{
This is an internal function, it will be excluded in future versions.
!!!!! DA IMPLEMENTARE !!!!!
}
