% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SMC_semiparBOOT_Cpp}
\alias{SMC_semiparBOOT_Cpp}
\title{Semiparametric bootstrap version of the Sun-McCabe score test.}
\usage{
SMC_semiparBOOT_Cpp(x, B, method)
}
\arguments{
\item{x}{NumericVector}

\item{B}{int}

\item{method}{unsigned int}
}
\description{
Semiparametric bootstrap version of the Sun-McCabe score test.
}
\details{
This is an internal function, it will be excluded in future versions.
}
