% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{SMCboot.test}
\alias{SMCboot.test}
\title{Sun McCabe score statistic to test for dependence in an integer autoregressive process}
\usage{
SMCboot.test(x, order = 1, arrival = "poisson", procedure, B = 499)
}
\arguments{
\item{x}{data vector.}

\item{order}{p, the order of the INAR(p) process.}

\item{arrival}{distribution of the innovation process.}

\item{procedure}{the bootstrap procedure to use, either "semiparametric" or "parmetric" .}

\item{B}{number of Bootstrap replications.}
}
\value{
A list with class "htest" containing the following components:
\itemize{
\item statistic the value of the bootstrap Score test statistic.
\item p.value the p-value for the test.
\item null.value the specified hypothesized value of the thinning operator under the null hypothesis.
\item alternative a character string describing the alternative hypothesis.
\item method a character string indicating what type of bootstrap Score test was performed.
\item data.name a character string giving the assumption about the dgp of the data.
\item estimate the value of the Score test statistic.
}
}
\description{
Sun McCabe score statistic to test for dependence in an integer autoregressive process
}
\details{
This is a one-sided test, where the null hypothesis is \eqn{\alpha = 0} and under the alternative hypothesis the true thinning operator is greater than 0. The observed Score test statistic is computed according to the arrival distribution, see \insertCite{sun2013score;textual}{INAr} for additional details, while the bootstrap is performed according to \insertCite{palazzo2022semiparametric;textual}{INAr}.
}
\references{
\insertAllCited{}
}
