% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SMCtest_boot}
\alias{SMCtest_boot}
\title{Wrapper function for compution the Sun-McCabe bootstrap score test.}
\usage{
SMCtest_boot(X, arrival, type, B)
}
\arguments{
\item{B}{int}

\item{x}{NumericVector}
}
\description{
Wrapper function for compution the Sun-McCabe bootstrap score test.
}
\details{
This is an internal function, it will be excluded in future versions.
}
