% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sortunique}
\alias{sortunique}
\title{Sort and remove duplicates from a numeric vector}
\usage{
sortunique(v)
}
\arguments{
\item{v}{NumericVector}
}
\description{
Sort and remove duplicates from a numeric vector
}
\details{
This is an internal function, it will be excluded in future versions.
}
