% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sunMC_Cpp}
\alias{sunMC_Cpp}
\title{Sun McCabe score statistic to test for dependence in an integer autoregressive process}
\usage{
sunMC_Cpp(x, method)
}
\arguments{
\item{x}{NumericVector}

\item{method}{unsigned int}
}
\description{
Sun McCabe score statistic to test for dependence in an integer autoregressive process
}
\details{
This is an internal function, it will be excluded in future versions.
}
