% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sunMC_parBOOT_Cpp}
\alias{sunMC_parBOOT_Cpp}
\title{Parametric bootstrap version of the Sun McCabe score test.}
\usage{
sunMC_parBOOT_Cpp(x, B, method)
}
\arguments{
\item{x}{NumericVector}

\item{B}{int}

\item{method}{unsigned int}
}
\description{
Parametric bootstrap version of the Sun McCabe score test.
}
\details{
This is an internal function, it will be excluded in future versions.
!!!!! DA IMPLEMENTARE !!!!!
}
