% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARtest.R
\name{ZIDItest}
\alias{ZIDItest}
\title{Joint test for zero-inflation and dispersion for the INAR(1) (Weiss et al, 2019)}
\usage{
ZIDItest(X, type = "pv")
}
\arguments{
\item{X}{vector, a series.}

\item{type}{character, one of the possible zero inflation tests, ... ("pv") or ... ("vdb").}
}
\value{
........
}
\description{
Joint test for zero-inflation and dispersion for the INAR(1) (Weiss et al, 2019)
}
\details{
The function performs the zero inflation test for the INAR(1) model.
}
\examples{
# ....... examples .....
# ZIDItest(X = rpois(100,2))

}
