% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARtest.R
\name{ZItest}
\alias{ZItest}
\title{Perform the zero inflation test for the INAR(1) (Weiss et al, 2019)}
\usage{
ZItest(X, type = "pv")
}
\arguments{
\item{X}{vector, a series.}

\item{type}{character, one of the possible zero inflation tests, ... ("pv") or ... ("vdb").}
}
\value{
........
}
\description{
Perform the zero inflation test for the INAR(1) (Weiss et al, 2019)
}
\details{
The function performs the zero inflation test for the INAR(1) model.
}
\examples{
# ....... examples .....
# ZItest(X = rpois(100,2))

}
