% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptosporidiosis.R
\docType{data}
\name{cryptosporidiosis}
\alias{cryptosporidiosis}
\title{Counts of cryptosporidiosis infect}
\format{
An object of class \code{"tibble"};
}
\source{
SurvStat@RKI 2.0 repository, \url{https://survstat.rki.de/}
}
\usage{
data(cryptosporidiosis)
}
\description{
This data contain a time series expressing the weekly counts of new infections in Germany (period 2002–2008).
}
\examples{
data(cryptosporidiosis)
\donttest{
plot(cryptosporidiosis$X,type="b")}
}
\references{
Robert-Koch-Institut (2016)

Example 5.1.7, Weiß, C. H. (2018). An introduction to discrete-valued time series. John Wiley & Sons.
}
\keyword{datasets}
