% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INAR1_different_innovations.R
\name{genINAR}
\alias{genINAR}
\title{Add together two numbers}
\usage{
genINAR(n, par, arrival = "poisson", burnout = 500)
}
\arguments{
\item{n}{integer, number of observations, length of the series.}

\item{par}{vector, parameters related with the model, see the details section.}

\item{arrival}{character, arrival distribution. Default value is \code{"poisson"}, alternative values are \code{"negbin"} for Negative Binomial, \code{"genpoi"} for Generalized Poisson and \code{"katz"} for the Katz family.}

\item{burnout}{integer, number of starting observations to discard, set to 500 by default.}
}
\value{
A number.
}
\description{
Add together two numbers
}
\details{
The function generates observations drawn from an INAR(1) model
\deqn{X_t = \alpha {*} X_{t-1} + \varepsilon_t}
where \eqn{*} is the binomial thinning operator and \eqn{\alpha} is the thinning parameter.
The first value of the \code{par} vector is the thinning operator \eqn{\alpha}, while the remaining depend mainly on the type of arrival distribution. The arrivals already implemented are listed below.

\strong{Poisson}. Innovations are drawn from a Poisson distribution \eqn{Poi(\lambda)}
\deqn{P(\varepsilon_t = k)= \dfrac{\lambda^k e^{-\lambda}}{k!},}
The positive real number \eqn{\lambda} is equal to the expected value and also to the variance.The parameter is then equal to:
\itemize{
\item \code{par} = c(\eqn{\alpha},\eqn{\lambda})
}

\strong{Negative Binomial}. Innovations are drawn from a Poisson distribution \eqn{Poi(\lambda)}
\itemize{
\item \code{par} = ...
}

\strong{Generalized Poisson}. Innovations are \eqn{ \varepsilon_t \sim Poi(\lambda)}
\itemize{
\item \code{par} = ...
}

\strong{Katz}. Innovations are \eqn{ \varepsilon_t \sim Poi(\lambda)}
\itemize{
\item \code{par} = ...
}

By the way, gatto gatto miao miao \code{\link[=genINAR]{genINAR()}}.
}
\examples{
# Generate 500 an INAR(1) model with thinning \alpha=0.5 Poisson(2) innovations
pars <- c(0.5,2)
INAr::genINAR(500, pars, arrival = "poisson")

}
