% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HMCtest_boot}
\alias{HMCtest_boot}
\title{Wrapper function for compution the Harris-McCabe bootstrap score test.
!!!WARNING!!! Still under development, do not use.}
\usage{
HMCtest_boot(X, B)
}
\arguments{
\item{X}{NumericVector}

\item{B}{int}
}
\description{
Wrapper function for compution the Harris-McCabe bootstrap score test.
!!!WARNING!!! Still under development, do not use.
}
\details{
This is an internal function, it will be excluded in future versions.
}
