% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARfit.R, R/genericfuns.R
\name{INAR}
\alias{INAR}
\alias{print.INAR}
\alias{summary.INAR}
\alias{plot.INAR}
\title{Fitting INAR(p) Models}
\usage{
INAR(X, order, arrival = "poisson", method = "CLS")

\method{print}{INAR}(x, digits = max(3L, getOption("digits") - 3L), se = TRUE, ...)

\method{summary}{INAR}(object, ...)

\method{plot}{INAR}(object)
}
\arguments{
\item{X}{data vector}

\item{order}{p, the order of the INAR(p) process}

\item{arrival}{distribution of the innovation process}

\item{method}{estimation method}

\item{x, }{an \code{INAR} object.}

\item{digits, }{number of digits.}

\item{se, }{whether to show standard error or not.}

\item{..., }{additional options}

\item{object, }{an \code{INAR} object}
}
\value{
The fitted model, an object of class \code{INAR}
}
\description{
summary method for class \code{INAR}.
}
\details{
Frontend function that estimates an INAR(p) model given the distribution of the innovation process.
}
