% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARfit.R, R/genericfuns.R
\name{INAR}
\alias{INAR}
\alias{print.INAR}
\alias{summary.INAR}
\alias{plot.INAR}
\alias{fitted.INAR}
\title{Fitting INAR(p) Models}
\usage{
INAR(X, p, inn = "poi", method = "CLS")

\method{print}{INAR}(x, digits = max(3L, getOption("digits") - 3L), se = TRUE, ...)

\method{summary}{INAR}(object, ...)

\method{plot}{INAR}(object)

\method{fitted}{INAR}(object, ...)
}
\arguments{
\item{X}{data vector}

\item{p}{the order of the INAR(p) process}

\item{inn}{distribution of the innovation process}

\item{method}{estimation method, one of "YW" (Yule-Walker), "CLS" (Conditional Least Squares), "CML" (Conditional Maximum Likelihood), "SP" (Saddlepoint Approximation)}

\item{x, }{an \code{INAR} object.}

\item{digits, }{number of digits.}

\item{se, }{whether to show standard error or not.}

\item{..., }{additional options}

\item{object, }{an \code{INAR} object}
}
\value{
The fitted model, an object of class \code{INAR}
}
\description{
summary method for class \code{INAR}.

summary method for class \code{INAR}.
}
\details{
This function estimates an INAR(p) model given the distribution of the innovation process.
}
