% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INARtest.R
\name{INARtest}
\alias{INARtest}
\title{Perform INAR tests}
\usage{
INARtest(X, type, B = 0)
}
\arguments{
\item{X}{vector, thinning parameters. The lenght of this vector defines the number of lags \code{p} of the INAR(p) process.}

\item{type}{character, the type of test to be performed. The options are "smc", "hmc" and "rho".}

\item{B}{integer, the number of bootstrap samples to be generated. The default value is 0, corresponding to no bootstrap samples.}
}
\value{
A number.
}
\description{
Perform INAR tests
}
\details{
The function performs tests for the INAR(1) model.
The tests are based on the Sun-McCabe (SMC), Harris-McCabe (HMC) and Rho statistics.
The tests can be performed using the original data or using bootstrap samples.
The bootstrap samples are generated using the parametric, semiparametric and nonparametric methods.
It is possible to perform the tests using the original data (B=0) or using bootstrap samples (B > 0).
}
\examples{
# ....... examples .....
# INARtest(X = rpois(100,2), type = "smc", B = 0)

}
