% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Katz.R
\name{pkatz}
\alias{pkatz}
\title{cumulative distribution function for the Katz distribution}
\usage{
pkatz(x, a, b)
}
\arguments{
\item{x}{A vector of non-negative integers representing the number of occurrences.}

\item{a}{A positive parameter of the Katz distribution.}

\item{b}{A positive parameter of the Katz distribution.}
}
\value{
A vector of cumulative probabilities corresponding to the input values.
}
\description{
This function computes the cumulative distribution function for the Katz distribution.
}
\examples{
katz_cdf(0:10, a = 2, b = 3)
}
