% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Katz.R
\name{qkatz}
\alias{qkatz}
\title{Quantile function for the Katz distribution}
\usage{
qkatz(p, a, b)
}
\arguments{
\item{p}{A vector of probabilities (between 0 and 1).}

\item{a}{A positive parameter of the Katz distribution.}

\item{b}{A positive parameter of the Katz distribution.}
}
\value{
A vector of quantiles corresponding to the input probabilities.
}
\description{
This function computes the quantile function for the Katz distribution.
}
\examples{
katz_quantile(c(0.1, 0.5, 0.9), a = 2, b = 3)
}
