% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Katz.R
\name{rkatz}
\alias{rkatz}
\title{random generation for the Katz distribution}
\usage{
rkatz(n, a, b)
}
\arguments{
\item{n}{The number of random samples to generate.}

\item{a}{A positive parameter of the Katz distribution.}

\item{b}{A positive parameter of the Katz distribution.}
}
\value{
A vector of random samples from the Katz distribution.
}
\description{
This function generates random samples from the Katz distribution.
}
\examples{
rkatz(10, a = 2, b = 0.5) # prima era b = 3 (check dominion parameters)
}
